implement `BasicBlockDataBuilder`
