+++ core/src/mast/serialization/mod.rs
@@ -1,14 +1,12 @@
 use alloc::{string::String, vec::Vec};
 use miden_crypto::{Felt, ZERO};
-use num_traits::ToBytes;
 use winter_utils::{
     ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable, SliceReader,
 };
 
 use crate::{
-    mast::{MerkleTreeNode, OperationOrDecorator},
-    AdviceInjector, AssemblyOp, DebugOptions, Decorator, DecoratorList, Operation, OperationData,
-    SignatureKind,
+    mast::MerkleTreeNode, AdviceInjector, AssemblyOp, DebugOptions, Decorator, DecoratorList,
+    Operation, OperationData, SignatureKind,
 };
 
 use super::{MastForest, MastNode, MastNodeId};
@@ -20,7 +18,7 @@ mod info;
 use info::{MastNodeInfo, MastNodeType};
 
 mod basic_block_data_builder;
-use basic_block_data_builder::StringTableBuilder;
+use basic_block_data_builder::BasicBlockDataBuilder;
 
 #[cfg(test)]
 mod tests;
@@ -71,8 +69,7 @@ impl Deserializable for StringRef {
 
 impl Serializable for MastForest {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        let mut string_table_builder = StringTableBuilder::new();
-        let mut data: Vec<u8> = Vec::new();
+        let mut basic_block_data_builder = BasicBlockDataBuilder::new();
 
         // magic & version
         target.write_bytes(MAGIC);
@@ -86,16 +83,19 @@ impl Serializable for MastForest {
 
         // MAST node infos
         for mast_node in &self.nodes {
-            let mast_node_info = mast_node_to_info(mast_node, &mut data, &mut string_table_builder);
+            let mast_node_info =
+                MastNodeInfo::new(mast_node, basic_block_data_builder.current_data_offset());
+
+            if let MastNode::Block(basic_block) = mast_node {
+                basic_block_data_builder.encode_basic_block(basic_block);
+            }
 
             mast_node_info.write_into(target);
         }
 
-        // strings table
-        let strings = string_table_builder.into_table(&mut data);
-        strings.write_into(target);
+        let (data, string_table) = basic_block_data_builder.into_parts();
 
-        // data blob
+        string_table.write_into(target);
         data.write_into(target);
     }
 }
@@ -158,223 +158,6 @@ impl Deserializable for MastForest {
         };
 
         Ok(mast_forest)
-    }
-}
-
-// TODOP: Make `MastNode` method (impl in this module)?
-fn mast_node_to_info(
-    mast_node: &MastNode,
-    data: &mut Vec<u8>,
-    string_table_builder: &mut StringTableBuilder,
-) -> MastNodeInfo {
-    use MastNode::*;
-
-    let ty = MastNodeType::new(mast_node);
-    let digest = mast_node.digest();
-
-    let offset = match mast_node {
-        Block(basic_block) => {
-            let offset: u32 = data
-                .len()
-                .try_into()
-                .expect("MastForest serialization: data field larger than 2^32 bytes");
-
-            for op_or_decorator in basic_block.iter() {
-                match op_or_decorator {
-                    OperationOrDecorator::Operation(operation) => encode_operation(operation, data),
-                    OperationOrDecorator::Decorator(decorator) => {
-                        encode_decorator(decorator, data, string_table_builder)
-                    }
-                }
-            }
-
-            offset
-        }
-        Join(_) | Split(_) | Loop(_) | Call(_) | Dyn | External(_) => 0,
-    };
-
-    MastNodeInfo { ty, offset, digest }
-}
-
-fn encode_operation(operation: &Operation, data: &mut Vec<u8>) {
-    data.push(operation.op_code());
-
-    // For operations that have extra data, encode it in `data`.
-    match operation {
-        Operation::Assert(value) | Operation::MpVerify(value) => {
-            data.extend_from_slice(&value.to_le_bytes())
-        }
-        Operation::U32assert2(value) | Operation::Push(value) => {
-            data.extend_from_slice(&value.as_int().to_le_bytes())
-        }
-        // Note: we explicitly write out all the operations so that whenever we make a modification
-        // to the `Operation` enum, we get a compile error here. This should help us remember to
-        // properly encode/decode each operation variant.
-        Operation::Noop
-        | Operation::FmpAdd
-        | Operation::FmpUpdate
-        | Operation::SDepth
-        | Operation::Caller
-        | Operation::Clk
-        | Operation::Join
-        | Operation::Split
-        | Operation::Loop
-        | Operation::Call
-        | Operation::Dyn
-        | Operation::SysCall
-        | Operation::Span
-        | Operation::End
-        | Operation::Repeat
-        | Operation::Respan
-        | Operation::Halt
-        | Operation::Add
-        | Operation::Neg
-        | Operation::Mul
-        | Operation::Inv
-        | Operation::Incr
-        | Operation::And
-        | Operation::Or
-        | Operation::Not
-        | Operation::Eq
-        | Operation::Eqz
-        | Operation::Expacc
-        | Operation::Ext2Mul
-        | Operation::U32split
-        | Operation::U32add
-        | Operation::U32add3
-        | Operation::U32sub
-        | Operation::U32mul
-        | Operation::U32madd
-        | Operation::U32div
-        | Operation::U32and
-        | Operation::U32xor
-        | Operation::Pad
-        | Operation::Drop
-        | Operation::Dup0
-        | Operation::Dup1
-        | Operation::Dup2
-        | Operation::Dup3
-        | Operation::Dup4
-        | Operation::Dup5
-        | Operation::Dup6
-        | Operation::Dup7
-        | Operation::Dup9
-        | Operation::Dup11
-        | Operation::Dup13
-        | Operation::Dup15
-        | Operation::Swap
-        | Operation::SwapW
-        | Operation::SwapW2
-        | Operation::SwapW3
-        | Operation::SwapDW
-        | Operation::MovUp2
-        | Operation::MovUp3
-        | Operation::MovUp4
-        | Operation::MovUp5
-        | Operation::MovUp6
-        | Operation::MovUp7
-        | Operation::MovUp8
-        | Operation::MovDn2
-        | Operation::MovDn3
-        | Operation::MovDn4
-        | Operation::MovDn5
-        | Operation::MovDn6
-        | Operation::MovDn7
-        | Operation::MovDn8
-        | Operation::CSwap
-        | Operation::CSwapW
-        | Operation::AdvPop
-        | Operation::AdvPopW
-        | Operation::MLoadW
-        | Operation::MStoreW
-        | Operation::MLoad
-        | Operation::MStore
-        | Operation::MStream
-        | Operation::Pipe
-        | Operation::HPerm
-        | Operation::MrUpdate
-        | Operation::FriE2F4
-        | Operation::RCombBase => (),
-    }
-}
-
-fn encode_decorator(
-    decorator: &Decorator,
-    data: &mut Vec<u8>,
-    string_table_builder: &mut StringTableBuilder,
-) {
-    // Set the first byte to the decorator discriminant.
-    //
-    // Note: the most significant bit is set to 1 (to differentiate decorators from operations).
-    {
-        let decorator_variant: EncodedDecoratorVariant = decorator.into();
-        data.push(decorator_variant.discriminant() | 0b1000_0000);
-    }
-
-    // For decorators that have extra data, encode it in `data` and `strings`.
-    match decorator {
-        Decorator::Advice(advice_injector) => match advice_injector {
-            AdviceInjector::MapValueToStack {
-                include_len,
-                key_offset,
-            } => {
-                data.write_bool(*include_len);
-                data.write_usize(*key_offset);
-            }
-            AdviceInjector::HdwordToMap { domain } => data.extend(domain.as_int().to_le_bytes()),
-
-            // Note: Since there is only 1 variant, we don't need to write any extra bytes.
-            AdviceInjector::SigToStack { kind } => match kind {
-                SignatureKind::RpoFalcon512 => (),
-            },
-            AdviceInjector::MerkleNodeMerge
-            | AdviceInjector::MerkleNodeToStack
-            | AdviceInjector::UpdateMerkleNode
-            | AdviceInjector::U64Div
-            | AdviceInjector::Ext2Inv
-            | AdviceInjector::Ext2Intt
-            | AdviceInjector::SmtGet
-            | AdviceInjector::SmtSet
-            | AdviceInjector::SmtPeek
-            | AdviceInjector::U32Clz
-            | AdviceInjector::U32Ctz
-            | AdviceInjector::U32Clo
-            | AdviceInjector::U32Cto
-            | AdviceInjector::ILog2
-            | AdviceInjector::MemToMap
-            | AdviceInjector::HpermToMap => (),
-        },
-        Decorator::AsmOp(assembly_op) => {
-            data.push(assembly_op.num_cycles());
-            data.write_bool(assembly_op.should_break());
-
-            // context name
-            {
-                let str_index_in_table =
-                    string_table_builder.add_string(assembly_op.context_name());
-                data.write_usize(str_index_in_table);
-            }
-
-            // op
-            {
-                let str_index_in_table = string_table_builder.add_string(assembly_op.op());
-                data.write_usize(str_index_in_table);
-            }
-        }
-        Decorator::Debug(debug_options) => match debug_options {
-            DebugOptions::StackTop(value) => data.push(*value),
-            DebugOptions::MemInterval(start, end) => {
-                data.extend(start.to_le_bytes());
-                data.extend(end.to_le_bytes());
-            }
-            DebugOptions::LocalInterval(start, second, end) => {
-                data.extend(start.to_le_bytes());
-                data.extend(second.to_le_bytes());
-                data.extend(end.to_le_bytes());
-            }
-            DebugOptions::StackAll | DebugOptions::MemAll => (),
-        },
-        Decorator::Event(value) | Decorator::Trace(value) => data.extend(value.to_le_bytes()),
     }
 }
